+++ src/raygun.rum.js
@@ -6,7 +6,7 @@
  * Licensed under the MIT license.
  */
 var raygunRumFactory = function (window, $, Raygun) {
-    Raygun.RealUserMonitoring = function (apiKey, apiUrl, makePostCorsRequest, user, version, excludedHostNames, excludedUserAgents, debugMode, maxVirtualPageDuration, ignoreUrlCasing) {
+    Raygun.RealUserMonitoring = function (apiKey, apiUrl, makePostCorsRequest, user, version, tags, excludedHostNames, excludedUserAgents, debugMode, maxVirtualPageDuration, ignoreUrlCasing, customTimingsEnabled) {
         var self = this;
         var _private = {};
 
@@ -18,6 +18,8 @@ var raygunRumFactory = function (window, $, Raygun) {
         this.excludedUserAgents = excludedUserAgents;
         this.maxVirtualPageDuration = maxVirtualPageDuration || 1800000; // 30 minutes
         this.ignoreUrlCasing = ignoreUrlCasing;
+        this.customTimingsEnabled = customTimingsEnabled;
+        this.pendingPerformancePayload = null;
 
         this.makePostCorsRequest = function (url, data) {
             if (self.excludedUserAgents instanceof Array) {
@@ -57,12 +59,13 @@ var raygunRumFactory = function (window, $, Raygun) {
         this.virtualPage = null;
         this.user = user;
         this.version = version;
+        this.tags = tags;
         this.heartBeatInterval = null;
         this.offset = 0;
 
         this.attach = function () {
             getSessionId(function (isNewSession) {
-                self.pageLoaded(isNewSession);
+              self.pageLoaded(isNewSession);
             });
 
             var clickHandler = function () {
@@ -87,6 +90,7 @@ var raygunRumFactory = function (window, $, Raygun) {
                             user: self.user,
                             version: self.version || 'Not supplied',
                             device: navigator.userAgent,
+                            tags: self.tags,
                             data: JSON.stringify(data)
                         }]
                     };
@@ -121,6 +125,7 @@ var raygunRumFactory = function (window, $, Raygun) {
                         type: 'session_start',
                         user: self.user,
                         version: self.version || 'Not supplied',
+                        tags: self.tags,
                         device: navigator.userAgent
                     }]
                 };
@@ -129,6 +134,7 @@ var raygunRumFactory = function (window, $, Raygun) {
             }
 
             self.sendPerformance(true, true);
+            
             self.heartBeat();
 
             if (typeof window.performance === 'object' && typeof window.performance.now === 'function') {
@@ -138,8 +144,39 @@ var raygunRumFactory = function (window, $, Raygun) {
             }
         };
 
+        this.sendCustomTimings = function (customTimings) {
+            if (typeof customTimings === 'object' && (
+              typeof customTimings.custom1 === 'number' ||
+              typeof customTimings.custom2 === 'number' ||
+              typeof customTimings.custom3 === 'number' ||
+              typeof customTimings.custom4 === 'number' ||
+              typeof customTimings.custom5 === 'number' ||
+              typeof customTimings.custom6 === 'number' ||
+              typeof customTimings.custom7 === 'number' ||
+              typeof customTimings.custom8 === 'number' ||
+              typeof customTimings.custom9 === 'number' ||
+              typeof customTimings.custom10 === 'number')) {
+                  if (self.pendingPerformancePayload) {
+                    var payloadObject = JSON.parse(self.pendingPerformancePayload);
+                    var resourceObjects = JSON.parse(payloadObject.eventData[0].data);
+                    
+                    resourceObjects[0].customTiming = customTimings;
+
+                    payloadObject.eventData[0].data = JSON.stringify(resourceObjects);
+
+                    self.makePostCorsRequest(self.apiUrl + '/events?apikey=' + encodeURIComponent(self.apiKey), JSON.stringify(payloadObject));
+
+                    self.pendingPerformancePayload = null;
+                  }
+              }
+        };
+
         this.setUser = function (user) {
             self.user = user;
+        };
+
+        this.withTags = function (tags) {
+            self.tags = tags;
         };
 
         this.endSession = function () {
@@ -174,6 +211,7 @@ var raygunRumFactory = function (window, $, Raygun) {
                                 user: self.user,
                                 version: self.version || 'Not supplied',
                                 device: navigator.userAgent,
+                                tags: self.tags,
                                 data: dataJson
                             }]
                         };
@@ -231,11 +269,17 @@ var raygunRumFactory = function (window, $, Raygun) {
                     user: self.user,
                     version: self.version || 'Not supplied',
                     device: navigator.userAgent,
+                    tags: self.tags,
                     data: JSON.stringify(performanceData)
                 }]
             };
 
-            self.makePostCorsRequest(self.apiUrl + '/events?apikey=' + encodeURIComponent(self.apiKey), JSON.stringify(payload));
+            if (!self.customTimingsEnabled) {
+              self.makePostCorsRequest(self.apiUrl + '/events?apikey=' + encodeURIComponent(self.apiKey), JSON.stringify(payload));
+            } else {
+              // Queue the WRT until the custom timings are provided
+              self.pendingPerformancePayload = JSON.stringify(payload);
+            }
         };
 
         function stringToByteLength(str) {
@@ -266,7 +310,7 @@ var raygunRumFactory = function (window, $, Raygun) {
                 var sessionCookie = readCookie(self.cookieName);
                 var id = readSessionCookieElement(sessionCookie, 'id');
 
-                if (id === 'undefined') {
+                if (id === 'undefined' || id === 'null') {
                     self.sessionId = randomKey(32);
                     createCookie(self.cookieName, self.sessionId);
                     callback(true);